+++ Core/XmppCore.cs
@@ -241,6 +241,7 @@ namespace S22.Xmpp.Core {
 			private set;
 		}
 
+
 		/// <summary>
 		/// The event that is raised when an unrecoverable error condition occurs.
 		/// </summary>